backend/requirements.txt
# BusFeed - Sistema de Transporte Público do DF
# Dependências básicas para Django com suporte geográfico

# Framework principal
Django==4.2.7
djangorestframework==3.14.0

# Banco de dados geográfico
psycopg2-binary==2.9.7
django-environ==0.11.2

# Suporte geográfico PostGIS
GDAL==3.7.3
Pillow==10.1.0

# Documentação automática da API
drf-spectacular==0.26.5

# CORS para comunicação com frontend React
django-cors-headers==4.3.1

# Servidor WSGI para produção
gunicorn==21.2.0

# Cache e otimização
redis==5.0.1
django-redis==5.4.0

# Monitoramento e logging
sentry-sdk[django]==1.38.0

# Utilitários para produção
whitenoise==6.6.0

# Utilitários para desenvolvimento (comentar em produção)
django-debug-toolbar==4.2.0 